ref(compiler): extract NumericOperationSpecialization from CallSpecialization
Final step of the CallSpecialization god-class split. Move the typed
numeric / comparison / equality operator family — typedBinaryOpName,
typedVariadicChain, the not-eq peephole, and the shared node→type-tag
inference (inferredTypeOfNode + its numeric recursion) plus the NUMERIC_*
/ EQUALITY_* tables — into a dedicated NumericOperationSpecialization.

inferredTypeOfNode is made public so the str-concat specialiser (which
stays in CallSpecialization) can reuse it — that shared helper was the
dependency that previously blocked this extraction.

Pure move: CallSpecialization dispatch and CallEmitter delegate to the new
class; the numeric unit tests move alongside it (CallSpecializationTest ->
NumericOperationSpecializationTest). Generated PHP is byte-identical (all
integration fixtures pass). CallSpecialization is now ~330 LOC, down from
1325.
